Small bit of changes
There are a lot of changes needed from here, but I at least simplified states slightly. This is a *huge* mess of inheritance, and you need to utilize interfaces more. The idle and move distinction is moot, and it's so hard to keep track of where certain variables are coming from because of the inheritance